change event registration to match services and mutations
