+++ src/lib/vocab/community/communityServices.ts
@@ -17,7 +17,7 @@ import {
 	ReadCommunities,
 	ReadCommunity,
 	UpdateCommunitySettings,
-} from '$lib/vocab/community/community.events';
+} from '$lib/vocab/community/communityEvents';
 import {toDefaultCommunitySettings} from '$lib/vocab/community/community.schema';
 
 const log = new Logger(gray('[') + blue('communityServices') + gray(']'));
